+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Added unit tests for SMTP authentication failure with username and AUTH LOGIN initial-response dispatch
 
 ### Changed